criando entidade do Usuario

+++ src/main/java/io/github/tfrazzao/rest/user/modelos/CriarUsuarioRequestBody.java
@@ -7,6 +7,6 @@ public class CriarUsuarioRequestBody {
 
     private String nome;
 
-    private Integer idade;
+    private Long idade;
 
 }
